CStructuedBuffer 구조 개선, TileMap 오류 해결.

+++ Project/Engine/Function.cpp
@@ -3,17 +3,17 @@
 
 void SaveStringToFile(const tstring& _str, FILE* _pFile)
 {
-	int iLen = (int)_str.length();
-	fwrite(&iLen, sizeof(int), 1, _pFile);
-	fwrite(_str.data(), sizeof(TCHAR), (size_t)iLen, _pFile);
+	size_t iLen = _str.length();
+	fwrite(&iLen, sizeof(size_t), 1, _pFile);
+	fwrite(_str.data(), sizeof(TCHAR), iLen, _pFile);
 }
 
 void LoadStringFromFile(tstring& _str, FILE* _pFile)
 {
 	TCHAR szBuffer[1024] = {};
-	int iLen = 0;
-	fread(&iLen, sizeof(int), 1, _pFile);
-	fread(szBuffer, sizeof(TCHAR), (size_t)iLen, _pFile);
+	size_t iLen = 0;
+	fread(&iLen, sizeof(size_t), 1, _pFile);
+	fread(szBuffer, sizeof(TCHAR), iLen, _pFile);
 	_str = szBuffer;
 }
